LLM/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
twilio==7.19.0
google-cloud-speech==2.16.1
google-cloud-translate==3.8.2
openai==0.27.5
faiss-cpu==1.7.3
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.2
anthropic==0.34.0
flask-cors==4.0.0
